created new class Launcher with main method, created FizzBuzz interface
