+++ FizzBuzz/src/main/java/rocks/kata/FizzBuzzLauncher.java
@@ -1,0 +1,7 @@
+package rocks.kata;
+
+public class FizzBuzzLauncher {
+    public static void main(String[] args) {
+        System.out.println("Hello launcher");
+    }
+}
